feat: laravel pint, ci and phpstan

+++ database/migrations/0000_01_01_000000_create_documents_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->enum('document_type', [DocumentTypeEnum::CNPJ, DocumentTypeEnum::CPF])
                 ->default('CPF');
-            $table->integer('number')->unique();
+            $table->bigInteger('number')->unique();
             $table->timestamps();
         });
     }
